# server/openapi.yaml
openapi: 3.1.0
info:
  title: OpenClaw Console API
  version: v1
  summary: Resource-oriented console API for OpenClaw model management.
  description: |
    AIP-style, resource-oriented HTTP API for configuring OpenClaw model settings
    and provider authentication.

    Mutating provider flows currently supported:
    - API Key providers flagged by `supportsApiKey=true`
    - OpenAI Codex OAuth subscription (`openai-codex`)

    Provider reads are discovered from local OpenClaw CLI.
servers:
  - url: http://127.0.0.1:18080
    description: Local default address
...
paths:
...
  /api/v1/modelSettings/default:
    get:
      tags: [ModelSettings]
      operationId: getDefaultModelSetting
      summary: Get default model setting
      responses:
        '200':
          description: Default model setting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelSetting'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [ModelSettings]
      operationId: patchDefaultModelSetting
      summary: Update default model setting
      description: Supports only `update_mask=default_model`.
      parameters:
        - name: update_mask
          in: query
          required: true
          schema:
            type: string
            enum: [default_model]
          description: Must be `default_model`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchModelSettingRequest'
      responses:
        '200':
          description: Updated default model setting
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModelSetting'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/providers:
    get:
      tags: [Providers]
      operationId: listProviders
      summary: List providers discovered by local OpenClaw CLI
      responses:
        '200':
          description: Provider list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListProvidersResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/providers/{provider}:
    get:
      tags: [Providers]
      operationId: getProvider
      summary: Get provider status
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
      responses:
        '200':
          description: Provider resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  '/api/v1/providers/{provider}:connectApiKey':
    post:
      tags: [Providers]
      operationId: connectProviderApiKey
      summary: Connect provider with API key
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConnectProviderAPIKeyRequest'
      responses:
        '200':
          description: Connection updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  '/api/v1/providers/{provider}:disconnect':
    post:
      tags: [Providers]
      operationId: disconnectProvider
      summary: Disconnect provider by removing all its auth profiles
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
      responses:
        '200':
          description: Provider after disconnect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Provider'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/auth:reset:
    post:
      tags: [Providers]
      operationId: resetAuth
      summary: Reset auth profiles with backup
      description: |
        Resets auth profiles in auth store and config files.
        Creates backups before write, then optionally restarts gateway.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetAuthRequest'
      responses:
        '200':
          description: Reset result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResetAuthResponse'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/providers/{provider}/authProfiles:
    get:
      tags: [AuthProfiles]
      operationId: listAuthProfiles
      summary: List auth profiles for provider
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
      responses:
        '200':
          description: List of auth profiles
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListAuthProfilesResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/providers/{provider}/authProfiles/{auth_profile}:
    get:
      tags: [AuthProfiles]
      operationId: getAuthProfile
      summary: Get auth profile
      parameters:
        - $ref: '#/components/parameters/ProviderParam'
        - $ref: '#/components/parameters/AuthProfileParam'
      responses:
        '200':
          description: Auth profile resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthProfile'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/channels/telegram:
    get:
      tags: [Channels]
      operationId: getTelegramChannel
      summary: Get Telegram channel configuration state
      responses:
        '200':
          description: Telegram channel resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramChannel'
        '500':
          $ref: '#/components/responses/InternalError'

    patch:
      tags: [Channels]
      operationId: patchTelegramChannel
      summary: Create or update Telegram channel configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchTelegramChannelRequest'
      responses:
        '200':
          description: Updated Telegram channel resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramChannel'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '500':
          $ref: '#/components/responses/InternalError'
  '/api/v1/channels/telegram:test':
    post:
      tags: [Channels]
      operationId: testTelegramChannel
      summary: Validate a Telegram bot token with the Telegram Bot API
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TestTelegramChannelRequest'
      responses:
        '200':
          description: Telegram bot token test result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramChannelTestResult'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '500':
          $ref: '#/components/responses/InternalError'
  '/api/v1/channels/telegram:disconnect':
    post:
      tags: [Channels]
      operationId: disconnectTelegramChannel
      summary: Remove Telegram channel configuration
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        '200':
          description: Telegram channel resource after disconnect
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TelegramChannel'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/modelCatalogEntries:
    get:
      tags: [ModelCatalogEntries]
      operationId: listModelCatalogEntries
      summary: List model catalog entries
      description: |
        Dual-mode endpoint:
        - with `provider`: paged catalog entries for that provider.
        - without `provider`: full available-model snapshot across providers.
      parameters:
        - name: provider
          in: query
          required: false
          schema:
            type: string
          description: Provider ID. If omitted, returns full available-model snapshot.
        - name: page_size
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
          description: Requested page size (provider mode only). Values >200 are clamped to 200 by the server.
        - name: page_token
          in: query
          required: false
          schema:
            type: string
          description: Opaque pagination token (provider mode only).
      responses:
        '200':
          description: Model catalog entries
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListModelCatalogEntriesResponse'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/codexAuthSessions:
    post:
      tags: [CodexAuthSessions]
      operationId: createCodexAuthSession
      summary: Create a codex auth session
      description: Starts an internal `openclaw onboard` flow in PTY and returns session state.
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCodexAuthSessionRequest'
      responses:
        '201':
          description: Created session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodexAuthSession'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/codexAuthSessions/{codex_auth_session}:
    get:
      tags: [CodexAuthSessions]
      operationId: getCodexAuthSession
      summary: Get codex auth session
      parameters:
        - $ref: '#/components/parameters/CodexAuthSessionParam'
      responses:
        '200':
          description: Session resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodexAuthSession'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  '/api/v1/codexAuthSessions/{codex_auth_session}:submitRedirect':
    post:
      tags: [CodexAuthSessions]
      operationId: submitCodexRedirect
      summary: Submit OAuth redirect URL for codex auth session
      parameters:
        - $ref: '#/components/parameters/CodexAuthSessionParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SubmitCodexRedirectRequest'
      responses:
        '200':
          description: Updated session resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodexAuthSession'
        '400':
          $ref: '#/components/responses/InvalidArgument'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/FailedPrecondition'
        '500':
          $ref: '#/components/responses/InternalError'
  '/api/v1/codexAuthSessions/{codex_auth_session}:cancel':
    post:
      tags: [CodexAuthSessions]
      operationId: cancelCodexAuthSession
      summary: Cancel codex auth session
      parameters:
        - $ref: '#/components/parameters/CodexAuthSessionParam'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
      responses:
        '200':
          description: Updated session resource
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CodexAuthSession'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  parameters:
    ProviderParam:
      name: provider
      in: path
      required: true
      schema:
        type: string
      description: Provider ID.
    AuthProfileParam:
      name: auth_profile
      in: path
      required: true
      schema:
        type: string
      description: Auth profile ID.
    CodexAuthSessionParam:
      name: codex_auth_session
      in: path
      required: true
      schema:
        type: string
      description: Codex auth session ID.
  responses:
    InvalidArgument:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: 400
              status: INVALID_ARGUMENT
              message: provider is required
    NotFound:
      description: Requested resource was not found.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: 404
              status: NOT_FOUND
              message: session not found
    FailedPrecondition:
      description: Resource state does not allow this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: 409
              status: FAILED_PRECONDITION
              message: session not ready for redirect
    InternalError:
      description: Internal server error.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: 500
              status: INTERNAL
              message: internal failure
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ErrorBody'
    ErrorBody:
      type: object
      required: [code, status, message]
      properties:
        code:
          type: integer
        status:
          type: string
        message:
          type: string
        details:
          type: array
          items: {}
    ModelSetting:
      type: object
      required: [name, defaultModel]
      properties:
        name:
          type: string
          example: modelSettings/default
        defaultModel:
          type: string
          example: openai-codex/gpt-5.3-codex
    PatchModelSettingRequest:
      type: object
      additionalProperties: false
      required: [defaultModel]
      properties:
        defaultModel:
          type: string
    PatchTelegramChannelRequest:
      type: object
      additionalProperties: false
      required: [enabled, dmPolicy, groupPolicy, requireMention]
      properties:
        enabled:
          type: boolean
        botToken:
          type: string
          description: Optional replacement bot token. Omit to keep the existing saved token.
        dmPolicy:
          type: string
          enum: [pairing, allowlist, open, disabled]
        allowFrom:
          type: array
          items:
            type: string
        groupPolicy:
          type: string
          enum: [allowlist, open, disabled]
        requireMention:
          type: boolean
    TestTelegramChannelRequest:
      type: object
      additionalProperties: false
      properties:
        botToken:
          type: string
          description: Optional unsaved bot token to validate. If omitted, the saved token is used.
    Provider:
      type: object
      required: [name, providerId, supportsApiKey, connection, authType, missingInUse]
      properties:
        name:
          type: string
          example: providers/openai
        providerId:
          type: string
        supportsApiKey:
          type: boolean
        connection:
          type: string
          enum: [NOT_CONFIGURED, CONNECTED]
        authType:
          type: string
          enum: [NONE, API_KEY, OAUTH, TOKEN]
        profileLabels:
          type: array
          items:
            type: string
        missingInUse:
          type: boolean
        oauthProviders:
          type: array
          items:
            type: string
    TelegramChannel:
      type: object
      required: [name, channelId, displayName, enabled, configured, mode, botTokenConfigured, dmPolicy, groupPolicy, requireMention, webhookUrlConfigured]
      properties:
        name:
          type: string
          example: channels/telegram
        channelId:
          type: string
          example: telegram
        displayName:
          type: string
          example: Telegram
        enabled:
          type: boolean
        configured:
          type: boolean
        mode:
          type: string
          enum: [polling, webhook]
        botTokenConfigured:
          type: boolean
        dmPolicy:
          type: string
          enum: [pairing, allowlist, open, disabled]
        allowFrom:
          type: array
          items:
            type: string
        groupPolicy:
          type: string
          enum: [allowlist, open, disabled]
        requireMention:
          type: boolean
        webhookUrlConfigured:
          type: boolean
        lastAppliedAction:
          type: string
    TelegramChannelTestResult:
      type: object
      required: [name, channelId, ok, message]
      properties:
        name:
          type: string
          example: channels/telegram:test
        channelId:
          type: string
          example: telegram
        ok:
          type: boolean
        message:
          type: string
        botId:
          type: integer
          format: int64
        botUsername:
          type: string
        botFirstName:
          type: string
    ProviderSummary:
      type: object
      required: [name, providerId, displayName, supportsApiKey, managed]
      properties:
        name:
          type: string
          example: providers/openai
        providerId:
          type: string
        displayName:
          type: string
          example: OpenAI
        supportsApiKey:
          type: boolean
        managed:
          type: boolean
    ListProvidersResponse:
      type: object
      required: [providers]
      properties:
        providers:
          type: array
          items:
            $ref: '#/components/schemas/ProviderSummary'
    ConnectProviderAPIKeyRequest:
      type: object
      additionalProperties: false
      required: [apiKey]
      properties:
        apiKey:
          type: string
          description: Provider API key.
    ResetAuthRequest:
      type: object
      additionalProperties: false
      properties:
        provider:
          type: string
          enum: [openai, openai-codex, all]
          default: openai
        restart:
          type: boolean
          default: true
    ResetAuthResponse:
      type: object
      required: [provider, restart, authStorePath, configPath, authProfilesRemoved, configProfilesRemoved, restarted, restartSkipped]
      properties:
        provider:
          type: string
          enum: [openai, openai-codex, all]
        restart:
          type: boolean
        authStorePath:
          type: string
        configPath:
          type: string
        authProfilesRemoved:
          type: array
          items:
            type: string
        configProfilesRemoved:
          type: array
          items:
            type: string
        authBackupPath:
          type: string
        configBackupPath:
          type: string
        restarted:
          type: boolean
        restartSkipped:
          type: boolean
        restartError:
          type: string
    AuthProfile:
      type: object
      required: [name, profileId, provider, type, status]
      properties:
        name:
          type: string
          example: providers/openai/authProfiles/openai:default
        profileId:
          type: string
        provider:
          type: string
          enum: [openai, openai-codex]
        type:
          type: string
          enum: [api_key, oauth, token]
        status:
          type: string
          enum: [STATIC, OK, EXPIRED, UNKNOWN]
        email:
          type: string
        expiresAt:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds.
    ListAuthProfilesResponse:
      type: object
      required: [authProfiles]
      properties:
        authProfiles:
          type: array
          items:
            $ref: '#/components/schemas/AuthProfile'
    ModelCatalogEntry:
      type: object
      required: [name, modelKey, displayName, provider, input, contextWindow, available]
      properties:
        name:
          type: string
          example: modelCatalogEntries/openai~gpt-5
        modelKey:
          type: string
        displayName:
          type: string
        provider:
          type: string
        input:
          type: string
        contextWindow:
          type: integer
          format: int64
        available:
          type: boolean
        tags:
          type: array
          items:
            type: string
    ListModelCatalogEntriesResponse:
      type: object
      required: [modelCatalogEntries]
      properties:
        modelCatalogEntries:
          type: array
          items:
            $ref: '#/components/schemas/ModelCatalogEntry'
        nextPageToken:
          type: string
    CreateCodexAuthSessionRequest:
      type: object
      additionalProperties: false
      properties:
        defaultModelHint:
          type: string
          description: Optional model to set after successful OAuth merge.
    SubmitCodexRedirectRequest:
      type: object
      additionalProperties: false
      required: [redirectUrl]
      properties:
        redirectUrl:
          type: string
          format: uri
          description: Must start with `http://localhost:1455/auth/callback`.
    CodexAuthSession:
      type: object
      required: [name, sessionId, state, expiresAt, createdAt]
      properties:
        name:
          type: string
          example: codexAuthSessions/0f4cd18f6fc1c3d487f30bc7
        sessionId:
          type: string
        state:
          type: string
          enum:
            - CREATED
            - LAUNCHING_ONBOARD
            - AWAITING_REDIRECT_URL
            - EXCHANGING_TOKEN
            - MERGING_CREDENTIALS
            - RESTARTING_SERVICE
            - SUCCEEDED
            - FAILED
            - CANCELLED
            - EXPIRED
        authUrl:
          type: string
          format: uri
        expiresAt:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds.
        createdAt:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds.
        defaultModelHint:
          type: string
        errorCode:
          type: string
        errorMessage:
          type: string